creates an encoding strategy defined by encoding property by config
