--- sanitizer/src/main/java/com/networknt/sanitizer/EncodingStrategy.java
@@ -1,9 +1,0 @@
-package com.networknt.sanitizer;
-
-public enum EncodingStrategy {
-    DEFAULT;
-
-    public static EncodingStrategy of(String value) {
-        return DEFAULT;
-    }
-}
