+++ sanitizer/src/main/java/com/networknt/sanitizer/enconding/DefaultEncoding.java
@@ -1,0 +1,11 @@
+package com.networknt.sanitizer.enconding;
+
+import org.owasp.encoder.Encode;
+
+public class DefaultEncoding implements Encoding {
+
+    @Override
+    public String apply(String data) {
+        return Encode.forJavaScriptSource(data);
+    }
+}
