+++ sanitizer/src/main/java/com/networknt/sanitizer/enconding/EncodingStrategy.java
@@ -1,0 +1,8 @@
+package com.networknt.sanitizer.enconding;
+
+public class EncodingStrategy {
+
+    public static Encoding of(String value) {
+        return new DefaultEncoding();
+    }
+}
