+++ sanitizer/src/test/java/com/networknt/sanitizer/EncodingStrategyTest.java
@@ -1,5 +1,7 @@
 package com.networknt.sanitizer;
 
+import com.networknt.sanitizer.enconding.DefaultEncoding;
+import com.networknt.sanitizer.enconding.EncodingStrategy;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -7,6 +9,6 @@ public class EncodingStrategyTest {
 
     @Test
     public void shouldGetDefaultStratetyIfDoesNotFindForValue() {
-        Assert.assertEquals(EncodingStrategy.DEFAULT, EncodingStrategy.of("anyString"));
+        Assert.assertTrue(EncodingStrategy.of("anyString") instanceof DefaultEncoding);
     }
 }
