+++ sanitizer/src/test/resources/config/sanitizer.yml
@@ -9,3 +9,5 @@ sanitizeBody: true
 
 # if it is enabled, does header need to be sanitized
 sanitizeHeader: true
+
+encoding: 'default'
